Albedo Bot/requirements.txt
aiohttp==3.6.2
anime-downloader==5.0.4
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
braceexpand==0.1.6
bs4==0.0.1
cachelib==0.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
cfgv==3.2.0
cfscrape==2.1.1
chardet==3.0.4
click==7.1.2
cloudant==2.13.0
colorama==0.4.3
coloredlogs==15.0
comtypes==1.1.7
cssselect==1.1.0
cycler==0.10.0
DateTime==4.3
dblpy==0.4.0
decorator==4.4.2
demjson==2.2.4
Deprecated==1.2.10
dhooks==1.1.4
discord==1.0.1
discord.py==1.5.0
disputils==0.1.3
distlib==0.3.1
Django==3.1.2
dnspython==2.0.0
ffmpeg==1.4
filelock==3.0.12
fire==0.3.1
Flask==1.1.2
fleep==1.0.1
futures==3.1.1
fuzzywuzzy==0.18.0
google-api-core==1.22.2
google-api-python-client==1.12.8
google-auth==1.22.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-core==1.4.4
google-cloud-translate==2.0.1
Google-Images-Search==1.3.4
googleapis-common-protos==1.52.0
googletrans==3.0.0
goslate==1.5.1
grpcio==1.34.0
gTTS==2.1.1
gTTS-token==1.1.3
h11==0.9.0
h2==3.2.0
howdoi==2.0.5
hpack==3.0.0
hstspreload==2020.8.8
httpcore==0.9.1
httplib2==0.18.1
httpx==0.13.3
humanfriendly==9.1
humanize==3.1.0
hyperframe==5.2.0
identify==1.5.10
idna==2.10
imgur-python==0.2.1
import-expression==1.1.4
iniconfig==1.1.1
ipykernel==5.3.4
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jupyter-client==6.1.7
jupyter-core==4.6.3
keep==2.9
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lo==0.2.0
log3==0.1.7
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
multidict==4.7.6
mysql-connector-python==8.0.22
nodeenv==1.5.0
nose==1.3.7
numpy==1.19.2
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.4.0.44
packaging==20.4
pandas==1.1.3
parso==0.7.1
pexels-api==1.0.1
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
pre-commit==2.9.2
prompt-toolkit==3.0.7
protobuf==3.13.0
psutil==5.7.2
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.9
PyDictionary==2.0.1
PyDrive==1.3.1
pyfiglet==0.8.post1
PyGithub==1.53
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
pymongo==3.11.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyquery==1.4.1
pyreadline==2.1
pySmartDL==1.3.4
PySSL==0.0.7
pytest==6.1.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-resize-image==1.1.19
pyttsx3==2.90
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
ratelimiter==1.2.0.post0
requests==2.24.0
requests-cache==0.5.2
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.6
simplejson==3.17.2
six==1.15.0
sniffio==1.1.0
soupsieve==2.0.1
sqlparse==0.3.1
tabulate==0.8.7
termcolor==1.1.0
terminaltables==3.1.0
toml==0.10.1
tornado==6.0.4
tox==3.20.1
translate==3.5.0
typing-extensions==3.7.4.2
unsplash-search==0.2.0
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.2.1
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
wikipedia==1.4.0
wrapt==1.12.1
xlrd==1.2.0
xmltodict==0.12.0
yarl==1.5.0
youtube-dl==2020.9.20
zope.interface==5.2.0
